Blade: reconstruct B_WinApp::GetWindow

+++ Blade/App.h
@@ -44,7 +44,7 @@ public:
     virtual void CallCdCb(int arg);
     virtual void ExitWithError(char *title, char* message) = 0;
     virtual void PrintWarning(const char *Title, const char *Message) = 0;
-    virtual int GetWindow() = 0;
+    virtual size_t GetWindow() = 0;
     virtual int GetModule() = 0;
     virtual int SetDirMap(const char *map) = 0;
     int LoadWorld(const char *file_name);
